refactoring and added hijri date api
