+++ src/types/api/index.ts
@@ -1,0 +1,3 @@
+export * from './aladhan';
+export * from './geoapify';
+export * from './openweather';
