+++ src/types/api/openweather.ts
@@ -1,0 +1,25 @@
+export type OpenWeatherForecastResponse = {
+  city: {
+    name: string;
+  };
+  list: Array<{
+    dt: number;
+    main: {
+      temp: number;
+      feels_like: number;
+      temp_min: number;
+      temp_max: number;
+      humidity: number;
+    };
+    weather: Array<{
+      id: number;
+      main: string;
+      description: string;
+      icon: string;
+    }>;
+    wind: {
+      speed: number;
+    };
+    dt_txt: string;
+  }>;
+};
